requirements.txt
numpy==1.18.4
paho-mqtt==1.5.0
pyserial==3.4
